move version speicifc func to version file
